refac: rename userApi to mainApi
